refactor: modularize size reporting and trim Apple runner
