+++ .github/workflows/size.yml
@@ -40,8 +40,8 @@ jobs:
       - name: Setup toolchain
         uses: ./.github/actions/setup-node-pnpm
 
-      - name: Preserve report script
-        run: cp scripts/size-report.mjs /tmp/agent-device-size-report.mjs
+      - name: Preserve report scripts
+        run: cp scripts/size-report*.mjs /tmp/
 
       # dist is fully determined by the base commit, so reuse it across PR runs
       # against the same base. Startup medians are still measured fresh on this
